+++ delete_hysds_cluster.sh
@@ -9,4 +9,16 @@ ${BASE_PATH}/delete_mozart.sh
 ${BASE_PATH}/delete_grq.sh
 
 # delete global configmap
-kubectl delete configmap hysds-global-config
+kubectl get configmap hysds-global-config 2>/dev/null
+if [ $? -eq 0 ]; then
+  kubectl delete configmap hysds-global-config
+fi
+
+# delete global secrets
+kubectl get secret hysds-global-secrets 2>/dev/null
+if [ $? -eq 0 ]; then
+  kubectl delete secret hysds-global-secrets
+fi
+
+# delete secrets
+rm -rf ${BASE_PATH}/secrets
